% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSettings.R
\name{gains_chart}
\alias{gains_chart}
\title{plotFunctions
Functions to plot identified charts}
\usage{
gains_chart(data, title, xlabel, ylabel)
}
\description{
plotFunctions
Functions to plot identified charts
}
