% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gainslift .R
\name{gainschart_diffsize}
\alias{gainschart_diffsize}
\title{gainslift gainschart_diffsize Function}
\usage{
gainschart_diffsize(gainsdata_diffsize, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}

\item{gainsdata.diffsize}{A gainsdata.diffsize object.
see \code{gainsdata.diffsize}.}
}
\description{
gainslift gainschart_diffsize Function
}
