% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainslift.R
\name{gainsliftTable}
\alias{gainsliftTable}
\title{gainslift gainsliftTable Function}
\usage{
gainsliftTable(predictions, responses, type = 'cum', cost = 0, benefit = 0)
}
\arguments{
\item{predictions}{A dataframe containing one or multiple numeric vectors of predicted response values.
All of the vectors must have the same length as \code{responses}.}

\item{responses}{a numeric vector of actual response values.}

\item{type}{'cum' will give a cumulative table.
'decile' will give a cumulative decile table.
'noncum' will give a non-cumulative decile table.
(The default is set as 'cum')}

\item{cost}{a numeric value of the cost of an incorrect classification.
e.g. If the cost of mailing a target is $5, then the cost should be set as 5.}

\item{benefit}{a numeric value of the benefit of a correct classification.
e.g. If the value of a responder is $20, then the benefit should be set as 25.}
}
\description{
gainslift gainsliftTable Function
}
