% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainslift.R
\name{gainsliftTable_diffsize}
\alias{gainsliftTable_diffsize}
\title{gainslift gainsliftTable_diffsize Function}
\usage{
gainsliftTable_diffsize(list, type = 'cum')
}
\arguments{
\item{list}{a list containing one or multiple lists of \code{list(predictions, reponses)}.
See \code{gainsliftTable} for the description of predict and reponse.}

\item{type}{'cum' will give a cumulative gains table.
'decile' will give a cumulative decile gains table.
'noncum' will give a non-cumulative decile gains table.
(The default is set as 'cum')}
}
\description{
gainslift gainsliftTable_diffsize Function
}
