% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gainslift .R
\name{liftchart}
\alias{liftchart}
\title{gainslift liftchart Function}
\usage{
liftchart(liftdata, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{liftdata}{A liftdata object.
see \code{liftdata}.}

\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}
}
\description{
gainslift liftchart Function
}
