% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsliftcharts .R
\name{liftchart_diffsize}
\alias{liftchart_diffsize}
\title{gainslift liftchart_diffsize Function}
\usage{
liftchart_diffsize(liftdata_diffsize, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}

\item{liftchart_diffsize}{A liftdata.diffsize object.
see \code{liftdata.diffsize}.}
}
\description{
gainslift liftchart_diffsize Function
}
