% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainslift.R
\name{liftdata.diffsize}
\alias{liftdata.diffsize}
\title{gainslift liftdata.diffsize Function}
\usage{
liftdata.diffsize(list, type = 'cum')
}
\arguments{
\item{list}{a list containing one or multiple lists of \code{list(predictions, reponses)}.
See \code{gainsdata} for the description of predict and reponse.}

\item{type}{'cum' will give a cumulative gains table.
'decile' will give a cumulative decile gains table.
'noncum' will give a non-cumulative decile gains table.
(The default is set as 'cum')}
}
\description{
gainslift liftdata.diffsize Function
}
