% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gainslift .R
\name{plot.gainsdata.diffsize}
\alias{plot.gainsdata.diffsize}
\title{gainslift plot.gainsdata.diffsize Function}
\usage{
plot.gainsdata.diffsize(gainsdata.diffsize, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{gainsdata.diffsize}{A gainsdata.diffsize object.
see \code{gainsdata.diffsize}.}

\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}
}
\description{
gainslift plot.gainsdata.diffsize Function
}
