% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gainslift .R
\name{plot.liftdata.diffsize}
\alias{plot.liftdata.diffsize}
\title{gainslift plot.liftdata.diffsize Function}
\usage{
plot.liftdata.diffsize(liftdata.diffsize, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{liftdata.diffsize}{A liftdata.diffsize object.
see \code{liftdata.diffsize}.}

\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}
}
\description{
gainslift plot.liftdata.diffsize Function
}
