% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainsliftcharts .R
\name{gainschart}
\alias{gainschart}
\title{gainslift gainschart Function}
\usage{
gainschart(gainsdata, title = NULL, xlabel = NULL, ylabel = NULL)
}
\arguments{
\item{title}{a text label for the main title.}

\item{xlabel}{a text label for the x axis.}

\item{ylabel}{a text label for the y axis.}

\item{gainsdata}{A gainsdata object.
see \code{gainsdata}.}
}
\description{
gainslift gainschart Function
}
